Move Session to librustc_session
