+++ src/librustc/lib.rs
@@ -64,7 +64,6 @@
 #![recursion_limit="512"]
 
 #[macro_use] extern crate bitflags;
-extern crate getopts;
 #[macro_use] extern crate scoped_tls;
 #[cfg(windows)]
 extern crate libc;
@@ -73,10 +72,6 @@ extern crate libc;
 #[macro_use] extern crate log;
 #[macro_use] extern crate syntax;
 #[macro_use] extern crate smallvec;
-
-// Use the test crate here so we depend on getopts through it. This allow tools to link to both
-// librustc_driver and libtest.
-extern crate test as _;
 
 #[cfg(test)]
 mod tests;
@@ -113,7 +108,7 @@ pub mod middle {
 }
 
 pub mod mir;
-pub mod session;
+pub use rustc_session as session;
 pub mod traits;
 pub mod ty;
