+++ src/librustc_session/code_stats.rs
